Application/requirements.txt
asgiref==3.8.1
Django==5.0.6
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
whitenoise
PyJWT